+++ src/components/common/sidebar/index.jsx
@@ -1,3 +1,4 @@
+// SideBar.jsx
 import React, { useEffect, useState } from 'react'
 import { Edit, Plus } from 'lucide-react'
 import { useDispatch, useSelector } from 'react-redux'
@@ -11,14 +12,13 @@ import AddMemberModal from '../../dashboard/add_member_to_workspace_modal'
 import CreateChannelModal from '../../dashboard/create_channel_modal'
 import CreateDMModal from '../../dashboard/create_dm_channel_modal'
 import WorkspaceDetailsModal from '../../dashboard/workspace_detail_modal'
+import UserCard from '../user_card'
 
 const SideBar = () => {
     const dispatch = useDispatch()
     const { channels, currentChannel } = useSelector((state) => state.channel)
     const { dms } = useSelector((state) => state.dm)
-
     const { currentWorkspace } = useSelector((state) => state.workspace)
-
     const { token, user } = useSelector((state) => state.auth)
 
     useEffect(() => {
@@ -50,119 +50,131 @@ const SideBar = () => {
     const handleShowCreateDMForm = () => {
         setShowCreateDMForm(!showCreateDMForm)
     }
+
+    const handleGoToWorkspaceHomePage = () => {
+        dispatch(setActiveChannel(null))
+    }
+
     return (
-        <div className="w-60 bg-[#3f0e40] text-white p-4">
-            <div className="flex justify-between items-center align-middle">
-                <h1 className="font-bold text-2xl self-center">
-                    {currentWorkspace.name}
-                </h1>
-                <div className="flex space-x-2">
-                    <button
-                        className="p-1 hover:bg-gray-100 rounded"
-                        onClick={handleShowAddMemberToWorkspaceForm}
+        <div className="w-60 bg-[#3f0e40] text-white flex flex-col justify-between h-full">
+            <div className="p-4">
+                <div className="flex justify-between items-center align-middle">
+                    <h1
+                        className="font-bold text-2xl self-center cursor-pointer"
+                        onClick={handleGoToWorkspaceHomePage}
                     >
-                        <Plus className="w-4 h-4 text-gray-500" />
-                    </button>
-                    <button
-                        className="p-1 hover:bg-gray-100 rounded"
-                        onClick={handleShowWorkspaceDetailsModal}
-                    >
-                        <Edit className="w-4 h-4 text-gray-500" />
-                    </button>
+                        {currentWorkspace.name}
+                    </h1>
+                    <div className="flex space-x-2">
+                        <button
+                            className="p-1 hover:bg-gray-100 rounded"
+                            onClick={handleShowAddMemberToWorkspaceForm}
+                        >
+                            <Plus className="w-4 h-4 text-gray-500" />
+                        </button>
+                        <button
+                            className="p-1 hover:bg-gray-100 rounded"
+                            onClick={handleShowWorkspaceDetailsModal}
+                        >
+                            <Edit className="w-4 h-4 text-gray-500" />
+                        </button>
+                    </div>
                 </div>
-            </div>
 
-            <hr className="my-5"></hr>
-            {/* Channels */}
-            <div className="mb-6 mt-4">
-                <div className="flex items-center justify-between mb-2">
-                    <h2 className="font-semibold">Channels</h2>
-                    <button
-                        className="p-1 hover:bg-gray-100 rounded"
-                        onClick={handleShowAddChannelForm}
-                    >
-                        <Plus className="w-4 h-4 text-gray-500" />
-                    </button>
-                </div>
-                {channels &&
-                    channels instanceof Array &&
-                    channels.map(
-                        (channel) =>
-                            channel.type !== 'dm' && (
-                                <div
-                                    key={channel._id}
-                                    className={`flex items-center justify-between p-2 cursor-pointer rounded ${
-                                        currentChannel._id === channel._id
-                                            ? 'bg-[#1164A3]'
-                                            : 'hover:bg-[#350d36]'
-                                    }`}
-                                    onClick={() =>
-                                        dispatch(setActiveChannel(channel))
-                                    }
-                                >
-                                    <span># {channel.name}</span>
-                                    {channel.unread > 0 && (
-                                        <span className="bg-red-500 rounded-full px-2 py-1 text-xs">
-                                            {channel.unread}
-                                        </span>
-                                    )}
-                                </div>
-                            )
-                    )}
-            </div>
-
-            {/* Direct Messages */}
-            <div>
-                <div className="flex items-center justify-between mb-2">
-                    <h2 className="font-semibold">Direct Messages</h2>
-                    <button
-                        className="p-1 hover:bg-gray-100 rounded"
-                        onClick={() => handleShowCreateDMForm()}
-                    >
-                        <Plus className="w-4 h-4 text-gray-500" />
-                    </button>
-                </div>
-                {dms &&
-                    dms instanceof Array &&
-                    dms.map(
-                        (channel) =>
-                            channel.type === 'dm' && (
-                                <div
-                                    key={channel._id}
-                                    className={`flex items-center gap-2 p-2 cursor-pointer rounded ${
-                                        currentChannel._id === channel._id
-                                            ? 'bg-[#1164A3]'
-                                            : 'hover:bg-[#350d36]'
-                                    }`}
-                                    onClick={() => {
-                                        dispatch(setActiveChannel(channel))
-                                        setShowUserProfile(true)
-                                    }}
-                                >
+                <hr className="my-5"></hr>
+                {/* Channels */}
+                <div className="mb-6 mt-4">
+                    <div className="flex items-center justify-between mb-2">
+                        <h2 className="font-semibold">Channels</h2>
+                        <button
+                            className="p-1 hover:bg-gray-100 rounded"
+                            onClick={handleShowAddChannelForm}
+                        >
+                            <Plus className="w-4 h-4 text-gray-500" />
+                        </button>
+                    </div>
+                    {channels &&
+                        channels instanceof Array &&
+                        channels.map(
+                            (channel) =>
+                                channel.type !== 'dm' && (
                                     <div
-                                        className={`w-2 h-2 rounded-full ${
-                                            channel.status === 'online'
-                                                ? 'bg-green-500'
-                                                : 'bg-gray-500'
+                                        key={channel._id}
+                                        className={`flex items-center justify-between p-2 cursor-pointer rounded ${
+                                            currentChannel?._id === channel._id
+                                                ? 'bg-[#1164A3]'
+                                                : 'hover:bg-[#350d36]'
                                         }`}
-                                    />
-                                    <NameToAvatar
-                                        name={
-                                            channel.members[0]._id !== user._id
-                                                ? channel.members[0].name
-                                                : channel.members[1].name
+                                        onClick={() =>
+                                            dispatch(setActiveChannel(channel))
                                         }
-                                        size={30}
-                                    />
-                                    <span>
-                                        {channel.members[0]._id !== user._id
-                                            ? channel.members[0].name
-                                            : channel.members[1].name}
-                                    </span>
-                                </div>
-                            )
-                    )}
+                                    >
+                                        <span># {channel.name}</span>
+                                        {channel.unread > 0 && (
+                                            <span className="bg-red-500 rounded-full px-2 py-1 text-xs">
+                                                {channel.unread}
+                                            </span>
+                                        )}
+                                    </div>
+                                )
+                        )}
+                </div>
+
+                {/* Direct Messages */}
+                <div>
+                    <div className="flex items-center justify-between mb-2">
+                        <h2 className="font-semibold">Direct Messages</h2>
+                        <button
+                            className="p-1 hover:bg-gray-100 rounded"
+                            onClick={() => handleShowCreateDMForm()}
+                        >
+                            <Plus className="w-4 h-4 text-gray-500" />
+                        </button>
+                    </div>
+                    {dms &&
+                        dms instanceof Array &&
+                        dms.map(
+                            (channel) =>
+                                channel.type === 'dm' && (
+                                    <div
+                                        key={channel._id}
+                                        className={`flex items-center gap-2 p-2 cursor-pointer rounded ${
+                                            currentChannel?._id === channel._id
+                                                ? 'bg-[#1164A3]'
+                                                : 'hover:bg-[#350d36]'
+                                        }`}
+                                        onClick={() => {
+                                            dispatch(setActiveChannel(channel))
+                                            setShowUserProfile(true)
+                                        }}
+                                    >
+                                        <div
+                                            className={`w-2 h-2 rounded-full ${
+                                                channel.status === 'online'
+                                                    ? 'bg-green-500'
+                                                    : 'bg-gray-500'
+                                            }`}
+                                        />
+                                        <NameToAvatar
+                                            name={
+                                                channel.members[0]._id !==
+                                                user._id
+                                                    ? channel.members[0].name
+                                                    : channel.members[1].name
+                                            }
+                                            size={30}
+                                        />
+                                        <span>
+                                            {channel.members[0]._id !== user._id
+                                                ? channel.members[0].name
+                                                : channel.members[1].name}
+                                        </span>
+                                    </div>
+                                )
+                        )}
+                </div>
             </div>
+            <UserCard />
             {/*Modal form to add members to workspace*/}
             <AddMemberModal
                 isOpen={showAddMemberToWorkspaceForm}
@@ -185,4 +197,5 @@ const SideBar = () => {
         </div>
     )
 }
+
 export default SideBar
