+++ src/pages/dashboard/index.jsx
@@ -1,14 +1,7 @@
-import React, { useEffect, useState } from 'react'
-import UserProfile from '../../components/common/user_profile'
 import Sidebar from '../../components/common/sidebar'
 import ChartArea from '../../components/chat_area'
-import { useDispatch, useSelector } from 'react-redux'
-import { fetchChannels } from '../../store/slices/channelSlice'
 
 const Dashboard = () => {
-    const [showUserProfile, setShowUserProfile] = useState(false)
-    const [selectedUser, setSelectedUser] = useState(null)
-
     return (
         <>
             <div className="h-screen flex flex-col">
@@ -17,11 +10,6 @@ const Dashboard = () => {
                     <ChartArea />
                 </div>
             </div>
-
-            {/* User Profile Modal */}
-            {showUserProfile && selectedUser && (
-                <UserProfile user={selectedUser} />
-            )}
         </>
     )
 }
